# docs/swagger.yaml
openapi: 3.0.3
info:
  title: SimpleShop
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
...
servers:
  - url: http://localhost:5005
...
paths:
...
  /orders:
    get:
      tags:
        - orders
      summary: Get All Orders
      operationId: getOrders
      responses:
        '200':
          description: successful receved Orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    post:
      tags:
        - orders
      summary: Create Order
      description: This can only be done by the logged in admin.
      operationId: createOrder
      requestBody:
        description: Created Order object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderDummy'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order dose already exist.
  /orders/{id}:
    get:
      tags:
        - orders
      summary: Get Order by ID
      description: This can only be done by the logged in admin.
      operationId: getOrder
      parameters:
        - name: id
          in: path
          description: public_id of Order that need to be updated
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order dose not exist.
        '405':
          description: Function not allowed
    put:
      tags:
        - orders
      summary: edit Order
      description: This can only be done by the logged in admin.
      operationId: editOrder
      parameters:
        - name: id
          in: path
          description: public_id of Order that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Created Order object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderDummy'
      responses:
        '200':
          description: successful deleted Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order dose not exist.
        '405':
          description: Function not allowed
    delete:
      tags:
        - orders
      summary: delete Order
      description: This can only be done by the logged in admin.
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          description: public_id of Order that need to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order dose not exist.
        '405':
          description: Function not allowed
  # ORDERPOS
  /orders/{id}/pos:
    get:
      tags:
        - orders
      summary: Get All OrderPOS of Order
      operationId: getOrderPoss
      responses:
        '200':
          description: successful receved Orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPOS'
    post:
      tags:
        - orders
      summary: Create OrderPos
      description: This can only be done by the logged in admin.
      operationId: createOrderPos
      requestBody:
        description: Created OrderPos object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderPOSDummy'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPOS'
        '400':
          description: Order dose already exist.
  /orders/{id}/pos/{pos_id}:
    get:
      tags:
        - orders
      summary: Get OrderPOS by ID
      description: This can only be done by the logged in admin.
      operationId: getOrderPOS
      parameters:
        - name: id
          in: path
          description: public_id of Order
          required: true
          schema:
            type: string
        - name: pos_id
          in: path
          description: public_id of OrderPOS 
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPOS'
        '404':
          description: Order dose not exist.
        '405':
          description: Function not allowed
    put:
      tags:
        - orders
      summary: edit OrderPOS
      description: This can only be done by the logged in admin.
      operationId: editOrderPOS
      parameters:
        - name: id
          in: path
          description: public_id of Order that need to be updated
          required: true
          schema:
            type: string
        - name: pos_id
          in: path
          description: public_id of OrderPOS 
          required: true
          schema:
            type: string
      requestBody:
        description: Created OrderPOS object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OrderPOSDummy'
      responses:
        '200':
          description: successful deleted Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPOS'
        '404':
          description: Order dose not exist.
        '405':
          description: Function not allowed
    delete:
      tags:
        - orders
      summary: delete Order
      description: This can only be done by the logged in admin.
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          description: public_id of Order that need to be deleted
          required: true
          schema:
            type: string
        - name: pos_id
          in: path
          description: public_id of OrderPOS 
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted OrderPOS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPOS'
        '404':
          description: Order dose not exist.
        '405':
          description: Function not allowed
components:
  schemas:
...
    Order:
      type: object
      properties:
        public_id:
          type: string
          example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
        supplier:
          type: string
          example: '522369cd-2185-442e-a88e-9dwetq9c2386'
    OrderDummy:
      type: object
      properties:
        supplier:
          type: string
          example: '522369cd-2185-442e-a88e-9dwetq9c2386'
        order_type:
          type: integer
          description: 0=supply 1=order
          example: 0
    # ORDERPOS
    OrderPOS:
      type: object
      properties:
        public_id:
          type: string
          example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
        order_id:
          type: string
          example: '522369cd-2185-442e-a88e-9dwetq9c2386'
        product_id:
          type: string
          example: '522369cd-2185-442e-a88e-9dwetq9c2386'
        costs:
          type: number
          example: 12.4
        amount:
          type: integer
          example: 12
    OrderPOSDummy:
      type: object
      properties:
        order_id:
          type: string
          example: '522369cd-2185-442e-a88e-9dwetq9c2386'
        product_id:
          type: string
          example: '522369cd-2185-442e-a88e-9dwetq9c2386'
        amount:
          type: integer
          example: 12
    # SUPPLIER
...